Long type for getRecommendedUsers

+++ FualMain/src/main/java/com/app/fual/FualMain/Services/ManagerService.java
@@ -311,7 +311,7 @@ public class ManagerService implements IManagerService{
 	}
 	
 	@Override
-	public List<Integer> getRecommendedUsers(Integer userId){
+	public List<Long> getRecommendedUsers(Long userId){
 	
 		return ((List<UserDataDTO>) iUserDataDAO.findTop10OrderByLevel()).stream().map(UserDataDTO::getId).collect(Collectors.toList());
